scanner/pyproject.toml
[project]
name = "adkflow-scanner"
version = "0.1.0"
description = "AI-powered codebase scanner for generating ADKFlow projects"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Mauro" }
]
keywords = ["adk", "google-adk", "agents", "workflow", "scanner"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[project.scripts]
adkflow-scanner = "scanner.cli:main"

[project.urls]
Homepage = "https://github.com/mauro/adkflow"
Repository = "https://github.com/mauro/adkflow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["**/*.py", "**/*.md", "pyproject.toml"]

[tool.hatch.build.targets.wheel]
packages = ["scanner"]

[tool.hatch.build.targets.wheel.sources]
"." = "scanner"

[tool.hatch.build.targets.wheel.force-include]
"__init__.py" = "scanner/__init__.py"
"agent.py" = "scanner/agent.py"
"callbacks.py" = "scanner/callbacks.py"
"cli.py" = "scanner/cli.py"
"orchestrator.py" = "scanner/orchestrator.py"
"agents" = "scanner/agents"
"layout" = "scanner/layout"
"prompts" = "scanner/prompts"
"schemas" = "scanner/schemas"
"tools" = "scanner/tools"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["scanner"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["."]
exclude = ["**/__pycache__"]
